chore: show paypal on boarding help message when modal close wihtou completing process
