+++ assets/src/js/admin/admin-settings.js
@@ -374,90 +374,3 @@ document.addEventListener( 'DOMContentLoaded', () => {
 		} );
 	} );
 } );
-
-// Handle paypal onboarding.
-document.addEventListener( 'DOMContentLoaded', () => {
-	const onBoardingButton = document.getElementById( 'js-give-paypal-on-boarding-handler' ),
-		disconnectPayPalAccountButton = document.getElementById( 'js-give-paypal-disconnect-paypal-account' ),
-		connectionSettingContainer = document.querySelector( '#give-paypal-commerce-account-manager-field-wrap .connection-setting' ),
-		disConnectionSettingContainer = document.querySelector( '#give-paypal-commerce-account-manager-field-wrap .disconnection-setting' ),
-		countryField = document.getElementById( 'paypal_commerce_account_country' );
-
-	/**
-	 * Remove error container.
-	 *
-	 * @since 2.9.6
-	 */
-	function removeErrors() {
-		const errorsContainer = document.querySelector( '.paypal-message-template' );
-
-		if ( errorsContainer ) {
-			errorsContainer.parentElement.parentElement.remove();
-		}
-	}
-
-	if ( onBoardingButton ) {
-		onBoardingButton.addEventListener( 'click', function( evt ) {
-			evt.preventDefault();
-			removeErrors();
-
-			const countryCode = countryField.value;
-			const buttonState = {
-				enable: () => {
-					onBoardingButton.disabled = false;
-					evt.target.innerText = onBoardingButton.getAttribute( 'data-initial-label' );
-				},
-				disable: () => {
-					// Preserve initial label.
-					if ( ! onBoardingButton.hasAttribute( 'data-initial-label' ) ) {
-						onBoardingButton.setAttribute( 'data-initial-label', onBoardingButton.innerText );
-					}
-
-					onBoardingButton.disabled = true;
-					evt.target.innerText = Give.fn.getGlobalVar( 'loader_translation' ).processing;
-				},
-			};
-
-			buttonState.disable();
-
-			fetch( ajaxurl + `?action=give_paypal_commerce_get_partner_url&countryCode=${ countryCode }` )
-				.then( response => response.json() )
-				.then( function( res ) {
-					if ( true === res.success ) {
-						const payPalLink = document.querySelector( '[data-paypal-button]' );
-
-						payPalLink.href = `${ res.data.partnerLink }&displayMode=minibrowser`;
-						payPalLink.click();
-					}
-
-					buttonState.enable();
-				}
-				);
-
-			return false;
-		} );
-	}
-
-	if ( disconnectPayPalAccountButton ) {
-		disconnectPayPalAccountButton.addEventListener( 'click', function( evt ) {
-			evt.preventDefault();
-			removeErrors();
-
-			new GiveConfirmModal( {
-				modalContent: {
-					title: givePayPalCommerce.translations.confirmPaypalAccountDisconnection,
-					desc: givePayPalCommerce.translations.disconnectPayPalAccount,
-				},
-				successConfirm: () => {
-					connectionSettingContainer.classList.remove( 'give-hidden' );
-					disConnectionSettingContainer.classList.add( 'give-hidden' );
-					countryField.parentElement.parentElement.classList.remove( 'hide-with-position' );
-
-					fetch( ajaxurl + '?action=give_paypal_commerce_disconnect_account' );
-				},
-			} ).render();
-
-			return false;
-		} );
-	}
-} );
